+++ package.json
@@ -13,6 +13,7 @@
     "axios": "^0.18.0",
     "material-design-icons-iconfont": "^3.0.3",
     "vue": "^2.5.11",
+    "vue-router": "^3.0.1",
     "vuetify": "^1.1.8",
     "vuex": "^3.0.1"
   },
@@ -30,7 +31,7 @@
     "cross-env": "^5.0.5",
     "css-loader": "^0.28.7",
     "file-loader": "^1.1.4",
-    "node-sass": "^4.5.3",
+    "node-sass": "^4.9.3",
     "sass-loader": "^6.0.6",
     "url-loader": "^1.0.1",
     "vue-loader": "^13.0.5",
