--- src/components/FooterComponent.vue
@@ -1,6 +1,0 @@
-<template>
-    <v-footer color="indigo" app>
-        <span class="white--text">&copy; 2018</span>
-    </v-footer>
-</template>
-
