--- src/components/SidebarComponent.vue
@@ -1,25 +1,0 @@
-<template>
-    <v-navigation-drawer
-        v-model="drawer"
-        fixed
-        app
-    >
-        <v-list dense>
-            <v-list-tile>
-                <v-list-tile-action>
-                    <v-icon>home</v-icon>
-                </v-list-tile-action>
-                <v-list-tile-content>
-                    <v-list-tile-title>Home</v-list-tile-title>
-                </v-list-tile-content>
-            </v-list-tile>
-        </v-list>
-      </v-navigation-drawer>
-</template>
-<script>
-export default {
-    props: {
-        drawer: Boolean
-    }
-}
-</script>
